requirements.txt
# requirements.txt - Ultimate Auto Trading System (Python 3.9+)

# Core Libraries (Python 3.9 compatible)
numpy>=1.22.0,<1.26.0
pandas>=1.5.0,<2.1.0
scikit-learn>=1.1.0,<1.4.0
scipy>=1.9.0,<1.12.0

# Configuration & Logging
PyYAML>=6.0
python-dotenv>=1.0.0

# Machine Learning - Traditional (Better Python 3.9 support)
xgboost>=1.7.0,<2.1.0
lightgbm>=4.0.0,<4.3.0
optuna>=3.0.0,<3.5.0

# Deep Learning (Full Python 3.9 support)
torch>=2.0.0,<2.2.0
torchvision>=0.15.0,<0.17.0
pytorch-lightning>=2.0.0,<2.2.0

# Reinforcement Learning (Python 3.9 native)
stable-baselines3>=2.0.0,<2.3.0
gymnasium>=0.28.0,<0.30.0
sb3-contrib>=2.0.0,<2.3.0

# Backtesting (Excellent Python 3.9 support)
vectorbt>=0.25.0,<0.26.0
backtrader>=1.9.78,<1.10.0
zipline-reloaded>=3.0.0  # Works better with Python 3.9

# Data Processing (Modern versions)
pandas-ta>=0.3.14b,<0.4.0
yfinance>=0.2.0,<0.3.0

# Visualization (Latest versions)
matplotlib>=3.6.0,<3.9.0
seaborn>=0.12.0,<0.14.0
plotly>=5.15.0,<5.18.0
bokeh>=3.0.0,<3.3.0

# Performance & Optimization (Python 3.9 optimized)
numba>=0.57.0,<0.59.0
joblib>=1.3.0,<1.4.0
cython>=3.0.0,<3.1.0

# Financial Data
alpha-vantage>=2.3.0,<2.4.0
yfinance>=0.2.0,<0.3.0

# Web & API (Modern async support)
requests>=2.28.0,<2.32.0
aiohttp>=3.8.0,<3.10.0
websockets>=11.0,<12.0

# Development Tools
jupyter>=1.0.0
ipykernel>=6.20.0
notebook>=6.5.0

# Testing (Latest versions)
pytest>=7.4.0,<8.0.0
pytest-asyncio>=0.21.0,<0.23.0
pytest-cov>=4.0.0,<5.0.0

# Code Quality (Python 3.9 compatible)
black>=23.0.0,<24.0.0
flake8>=6.0.0,<7.0.0
mypy>=1.5.0,<1.8.0

# Memory & Performance Monitoring
psutil>=5.9.0,<6.0.0
memory-profiler>=0.61.0,<0.62.0

# Time Series Analysis (Enhanced for Python 3.9)
statsmodels>=0.14.0,<0.15.0
arch>=6.0.0,<7.0.0
pmdarima>=2.0.0,<2.1.0

# Risk Management (Updated versions)
pyfolio>=0.9.2
empyrical>=0.5.5
quantstats>=0.0.62

# Alternative Data (Python 3.9 compatible)
beautifulsoup4>=4.11.0,<4.13.0
selenium>=4.10.0,<5.0.0

# Crypto APIs (Optional - Modern versions)
ccxt>=4.0.0,<5.0.0
python-binance>=1.0.19
web3>=6.8.0,<7.0.0

# Monitoring & Alerts (Latest)
prometheus-client>=0.17.0,<0.19.0
slack-sdk>=3.21.0,<3.23.0

# Configuration Management (Python 3.9 native)
hydra-core>=1.3.0,<1.4.0
omegaconf>=2.3.0,<2.4.0

# Experiment Tracking (Latest)
wandb>=0.15.0,<0.16.0
tensorboard>=2.13.0,<2.15.0
mlflow>=2.5.0,<2.8.0

# Production Deployment (Python 3.9 optimized)
uvicorn[standard]>=0.23.0,<0.25.0
gunicorn>=21.0.0,<22.0.0
fastapi>=0.100.0,<0.105.0

# GPU Acceleration (Optional)
cupy-cuda11x>=12.0.0; platform_machine=="x86_64"
cudf-cu11>=23.06.0; platform_machine=="x86_64"